requirements.txt
databricks-connect==13.3.0  # align this with the cluster version
pytest